KIGG advanced server running, profile generic not working

+++ development/include/serverevents.h
@@ -52,8 +52,8 @@ extern "C" {
   */
     extern unsigned char svr_isTarget(
         dlmsSettings* settings,
-        unsigned long serverAddress,
-        unsigned long clientAddress);
+        uint32_t serverAddress,
+        uint32_t clientAddress);
 
     /**
     * Get attribute access level.
